requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.1

# Security
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0

# Google AI
google-generativeai>=0.3.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0.post1
Werkzeug==3.0.1
